Added README... again
Okay so TL;DR, worked hard on Arrays Pt. 3 (did it in one day, because was lazy for the whole week), anyway, pushed it to master, which didn't merge, so I had to unpush, but that deleted basically everything from Arrays pt. 3, including every header file. Sooooo I had to go looking for those in Git Commit history and eventually I found it, but somehow managed to delete README in the process, so here we are now